修改goods_table中的count,packageNum字段为整型；修改SQL语句将对TEXT的编码设定为utf8;对充值历史和交易历史的表添加唯一的token字段和用户是否显示的字段；修改token的生成算法；修改部分api：增加字段和加入分页功能

+++ src/main/java/com/fenrir/server/api/AdminResource.java
@@ -47,9 +47,9 @@ public class AdminResource {
             @FormParam("barCode") String barCode,
             @FormParam("salePrice") Float salePrice,
             @FormParam("costPrice") Float costPrice,
-            @FormParam("count") Float count,
+            @FormParam("count") Integer count,
             @FormParam("unit") String unit,
-            @FormParam("packageNum") Float packageNum,
+            @FormParam("packageNum") Integer packageNum,
             @FormParam("note") String note,
             @FormParam("className") String className
     ) throws Exception {
@@ -80,9 +80,9 @@ public class AdminResource {
             @FormParam("barCode") String barCode,
             @FormParam("salePrice") Float salePrice,
             @FormParam("costPrice") Float costPrice,
-            @FormParam("count") Float count,
+            @FormParam("count") Integer count,
             @FormParam("unit") String unit,
-            @FormParam("packageNum") Float packageNum,
+            @FormParam("packageNum") Integer packageNum,
             @FormParam("note") String note,
             @FormParam("className") String className
     ) throws Exception {
