+++ src/main/java/com/fenrir/server/model/api/GoodsModel.java
@@ -12,9 +12,9 @@ public class GoodsModel {
     private String barCode;
     private Float salePrice;
     private Float costPrice;
-    private Float count;
+    private Integer count;
     private String unit;
-    private Float packageNum;
+    private Integer packageNum;
     private String note;
     private String className;
